Code arranged into component and container folders to keep it organised as project grows
