+++ src/components/Header.js
@@ -1,9 +1,6 @@
 import React, { Component } from 'react';
 // import logo from './logo.svg';
 
-import logo from './logo.svg';
-import './App.css';
-
 function Header() {
   return (
     <div className="Header">
