+++ src/components/__tests__/Header.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Header from '../Header';
+
+test('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Header />, div);
+});
