+++ src/containers/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import Header from './Header'
-import './App.css';
+import Header from '../components/Header'
+import '../App.css';
 
 function App() {
   return (
